backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
apscheduler==3.10.4
requests==2.32.3
gunicorn==22.0.0
greenlet==3.1.1
pyyaml==6.0.2
bcrypt==4.2.1
flask-session==0.8.0
reportlab==4.2.5
tzdata==2024.2
